+++ modules/dxp/apps/commerce/commerce-product-options-web/src/main/java/com/liferay/commerce/product/options/web/internal/portlet/action/EditCPOptionValueMVCActionCommand.java
@@ -114,7 +114,7 @@ public class EditCPOptionValueMVCActionCommand extends BaseMVCActionCommand {
 		String name = ParamUtil.getString(actionRequest, "name");
 		Map<Locale, String> titleMap = LocalizationUtil.getLocalizationMap(
 			actionRequest, "title");
-		int priority = ParamUtil.getInteger(actionRequest, "priority");
+		double priority = ParamUtil.getDouble(actionRequest, "priority");
 
 		ServiceContext serviceContext = ServiceContextFactory.getInstance(
 			CPOptionValue.class.getName(), actionRequest);
